feat(GAPP-1260): remove all @agm packages and references
