--- src/app/tools/gen-mapper/dialogs/location-dialog/location-dialog.component.html
@@ -1,47 +1,0 @@
-<div fxLayout="column"
-     style="flex: 1;">
-    <div class="search-box"
-         fxLayout="column">
-        <div class="search-control">
-            <input matInput
-                   placeholder="Search..."
-                   autocomplete="off"
-                   autocapitalize="off"
-                   spellcheck="off"
-                   type="text"
-                   class="form-control"
-                   #search
-                   [formControl]="searchControl" />
-
-            <mat-icon *ngIf="searchControl.value"
-                      (click)="clear()"
-                      class="clear-icon">clear</mat-icon>
-        </div>
-    </div>
-
-    <agm-map [latitude]="latitude"
-             [longitude]="longitude"
-             [zoom]="zoom"
-             [disableDefaultUI]="true"
-             [style.height]="height + 'px'"
-             (mapReady)="mapReadyHandler($event)">
-        <agm-marker [latitude]="markerLatitude"
-                    [longitude]="markerLongitude"
-                    [markerDraggable]="true"
-                    (dragEnd)="markerDragEnd($event)"></agm-marker>
-    </agm-map>
-</div>
-
-<div mat-dialog-actions
-     fxLayoutAlign="end">
-    <button mat-raised-button
-            color="accent"
-            (click)="onSubmit()">
-        {{'Common_Continue' | translate}}
-    </button>
-    <button mat-button
-            color="warn"
-            [mat-dialog-close]="false">
-        {{'Common_Cancel' | translate}}
-    </button>
-</div>
