--- src/app/tools/gen-mapper/dialogs/location-dialog/location-dialog.component.scss
@@ -1,40 +1,0 @@
-@import '~styles/mixins';
-
-:host {
-    @include dialog();
-    position: relative;
-    min-height: 280px;
-
-    agm-map {
-        margin-top: -24px;
-        margin-left: -24px;
-        margin-right: -24px;
-    }
-
-    .search-box {
-        position: absolute;
-        top: 0;
-        left: 0;
-        right: 0;
-        z-index: 2;
-
-        input {
-            padding: 10px;
-            border-radius: 5px;
-            line-height: 1;
-            background-color: white;
-            width: calc(100% - 20px);;
-            box-shadow: 0 3px 5px rgba(0,0,0,0.2);
-        }
-
-        .clear-icon {
-            position: absolute;
-            top: 0;
-            right: 0;
-            margin: 7px;
-            color: red;
-            cursor: pointer;
-            user-select: none;
-        }
-    }
-}
