--- src/app/tools/gen-mapper/dialogs/location-dialog/location-dialog.component.spec.ts
@@ -1,25 +1,0 @@
-import { ComponentFixture, TestBed, waitForAsync } from '@angular/core/testing';
-
-import { LocationDialogComponent } from './location-dialog.component';
-import { configureTestSuite } from 'ng-bullet';
-
-describe.skip('LocationDialogComponent', () => {
-    let component: LocationDialogComponent;
-    let fixture: ComponentFixture<LocationDialogComponent>;
-
-    configureTestSuite(() => {
-        TestBed.configureTestingModule({
-            declarations: [LocationDialogComponent]
-        });
-    });
-
-    beforeEach(() => {
-        fixture = TestBed.createComponent(LocationDialogComponent);
-        component = fixture.componentInstance;
-        fixture.detectChanges();
-    });
-
-    it('should create', () => {
-        expect(component).toBeTruthy();
-    });
-});
