+++ src/app/tools/gen-mapper/gen-mapper.module.ts
@@ -1,5 +1,3 @@
-import { AgmCoreModule } from '@agm/core';
-import { AgmMarkerClustererModule } from '@agm/markerclusterer';
 import { CommonModule } from '@angular/common';
 import { NgModule } from '@angular/core';
 import { ReactiveFormsModule } from '@angular/forms';
@@ -15,7 +13,6 @@ import { ConfirmDialogComponent } from './dialogs/confirm-dialog/confirm-dialog.
 import { CreateDocumentDialogComponent } from './dialogs/create-document-dialog/create-document-dialog.component';
 import { EditDocumentDialogComponent } from './dialogs/edit-document-dialog/edit-document-dialog.component';
 import { InvalidCsvDialogComponent } from './dialogs/invalid-csv-dialog/invalid-csv-dialog.component';
-import { LocationDialogComponent } from './dialogs/location-dialog/location-dialog.component';
 import { PeopleDialogComponent } from './dialogs/people-dialog/people-dialog.component';
 import { CountryPickerComponent } from './dialogs/people-group-dialog/country-picker/country-picker.component';
 import { PeopleGroupDialogComponent } from './dialogs/people-group-dialog/people-group-dialog.component';
@@ -54,11 +51,6 @@ import { LocationMapDialogComponent } from './dialogs/location-map-dialog/locati
         ReactiveFormsModule,
         RouterModule,
         NgxChartsModule,
-        AgmCoreModule.forRoot({
-            apiKey: environment.apiKey,
-            libraries: ['places']
-        }),
-        AgmMarkerClustererModule,
         TranslateModule,
     ],
     declarations: [
@@ -71,7 +63,6 @@ import { LocationMapDialogComponent } from './dialogs/location-map-dialog/locati
         DocumentNameControlComponent,
         DocumentsSidenavComponent,
         CreateDocumentDialogComponent,
-        LocationDialogComponent,
         NodeDrawerComponent,
         PeopleGroupDialogComponent,
         CountryPickerComponent,
@@ -116,7 +107,6 @@ import { LocationMapDialogComponent } from './dialogs/location-map-dialog/locati
     entryComponents: [
         ConfirmDialogComponent,
         CreateDocumentDialogComponent,
-        LocationDialogComponent,
         PeopleGroupDialogComponent,
         InvalidCsvDialogComponent,
         SavingSnackbarComponent,
